requirements.txt
absl-py==0.13.0
attrs==21.2.0
certifi==2020.6.20
cffi @ file:///tmp/build/80754af9/cffi_1600699165083/work
chardet==3.0.4
colorama==0.3.9
cryptography @ file:///tmp/build/80754af9/cryptography_1601047419573/work
cvzone==1.5.0
cycler==0.10.0
Cython==0.29.4
dataclasses==0.8
fire==0.1.3
h5py==2.8.0
idna==2.8
kiwisolver==1.2.0
llvmlite==0.32.1
lxml==4.5.2
matplotlib==2.2.3
mediapipe==0.8.3
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
numba==0.39.0
numpy==1.19.3
opencv-python==3.4.3.18
pandas==0.23.4
Pillow==7.2.0
protobuf==3.13.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytils==0.3
pytz==2020.1
requests==2.21.0
scipy==1.1.0
six==1.15.0
tensorboardX==1.6
tools==0.1.9
torch @ file:///media/yyy/9B5BAA2D94B22124/SiamMask/torch-0.4.1-cp36-cp36m-manylinux1_x86_64.whl
torchvision==0.2.1
tqdm==4.29.1
urllib3==1.24.3
